+++ test/test_browser.py
@@ -2675,6 +2675,9 @@ Module["preRun"] = () => {
       self.cflags.append('--pre-js=pre.js')
     self.btest_exit('test_html5_core.c', cflags=opts)
 
+  def test_html5_remove_event_listener(self):
+    self.btest_exit('test_html5_remove_event_listener.c')
+
   @parameterized({
     '': ([],),
     'closure': (['-O2', '-g1', '--closure=1'],),